Safari-609.1.11.1/LayoutTests/fast/history/page-cache-geolocation-active-watcher-expected.txt

Tests that a page actively using geolocation can go into the page cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
Watcher callback called
PASS testRunner.isGeolocationProviderActive() is true
PASS stage == 'in_page_cache' is false
PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy

pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
PASS testRunner.isGeolocationProviderActive() is false
Watcher callback called
PASS testRunner.isGeolocationProviderActive() is true
PASS stage == 'in_page_cache' is false
PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy

PASS successfullyParsed is true

TEST COMPLETE
